Add file utils for utils methods
