+++ src/carowner.cpp
@@ -1,7 +1,7 @@
 #include "carowner.h"
 #include <iostream>
 
-CarOwner::CarOwner()
+CarOwner::CarOwner(std::string name, std::string _surname) : _name{name}, _surname{_surname}
 {
 }
 
@@ -12,4 +12,15 @@ void CarOwner::setCarForOwner(CarBrand carBrand, CarBodyStyle CarBodyStyle)
 
 void CarOwner::setAmountCash(int amountCash)
 {
+    _amountCash = amountCash;
+}
+
+const std::string CarOwner::getName() const 
+{
+    return _name;
+}
+
+const std::string CarOwner::getSurname() const 
+{
+    return _surname;
 }
