+++ src/carowner.h
@@ -3,15 +3,26 @@
 #include "icar.h"
 #include "car.h"
 
+#include <string>
+
 class CarOwner
 {
 public:
-  CarOwner();
+  CarOwner(std::string name, std::string _surname);
 
-  void setCarForOwner(CarBrand carBrand, CarBodyStyle CarBodyStyle);
   void setAmountCash(int amountCash);
+  void setCarForOwner(CarBrand carBrand, CarBodyStyle CarBodyStyle);
+
+  const Car *getOwnerCar() const;
+  const float getAmountCash() const;
+  const std::string getName() const;
+  const std::string getSurname() const;
+
+
 
 private:
-  Car *_ownerCar;
   float _amountCash;
+  Car *_ownerCar{nullptr};
+  std::string _name{"unknow"};
+  std::string _surname{"unknow"};
 };
